Isolate test examples
Isolate the examples by running them one at a time, and then unwinding the stack after each one. This is implemented as a search through the describe-block tree to find the current example to be run. After running each example, look for the next one to run. Rinse and repeat until all examples have been run.